sources.py: fix unconditional return in map value clone
In the generated C code for cloning the object values of a
mapStringObject, an 'if' statement checking for cloning failure
was followed by an erroneous semicolon.

This caused the 'if' to have an empty body, leading to the
subsequent 'return NULL;' being executed unconditionally in
the first iteration of the value-cloning loop if the map
was not empty. Consequently, cloning a mapStringObject with
one or more elements would always fail prematurely when this
specific path for object-values was taken.

Removing the semicolon makes the 'return NULL;' conditional
upon actual cloning failure, as intended.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>